requirements.txt
flask
flask-socketio
eventlet
python-dotenv
gunicorn